test: 테스트 수정 및 제거
