--- ai/tests/local/__init__.py
@@ -1,4 +1,0 @@
-"""
-로컬 환경 테스트 패키지 입니다.
-CI/CD 과정에서 계속 LLM을 호출할 수 없기 때문에 테스트를 나눕니다.
-"""
