--- ai/tests/unit/searcher/analyze_message_test.py
@@ -1,54 +1,0 @@
-import pytest
-import asyncio
-import json
-from unittest.mock import AsyncMock, patch
-from ai.services.openai_service import OpenAIService
-from ai.templateEngine.searcher import Searcher
-
-
-def test_analyze_message_with_mock():
-    service = OpenAIService()
-    searcher = Searcher(service)
-
-    # classify_category_and_type 가짜 응답
-    fake_classify_response = {
-        "category_sub": "주문 완료 안내",
-        "type": "BASIC",
-        "has_channel_link": False,
-        "has_extra_info": False,
-        "explain_category_sub": "Mock explanation for category",
-        "explain_type": "Mock explanation for type"
-    }
-
-    # extract_message_fields 가짜 응답
-    fake_extract_response = {
-        "intent_type": "정보성",
-        "recipient_scope": "구매자",
-        "links_allowed": True,
-        "variables": ["#{수신자명}", "#{주문번호}", "#{주문일시}", "#{주문상품}"]
-    }
-
-    # 두 메서드 동시에 Mock 처리
-    with patch.object(searcher, "classify_category_and_type", new=AsyncMock(return_value=fake_classify_response)), \
-         patch.object(searcher, "extract_message_fields", new=AsyncMock(return_value=fake_extract_response)):
-
-        result = asyncio.run(
-            searcher.analyze_message(
-                "안녕하세요 #{수신자명}님, 주문이 완료되었습니다.",
-                "주문/결제",
-                ["주문 완료 안내", "결제 완료 안내", "배송 준비 안내"]
-            )
-        )
-
-    # 검증: 두 결과가 합쳐졌는지 확인
-    assert isinstance(result, dict)
-
-    # classify 결과 확인
-    assert result["category_sub"] == "주문 완료 안내"
-    assert result["type"] == "BASIC"
-
-    # extract 결과 확인
-    assert result["intent_type"] == "정보성"
-    assert result["recipient_scope"] == "구매자"
-    assert result["links_allowed"] is True
-    assert "#{주문번호}" in result["variables"]
